Overlays dismiss. Connsvc fix. Some bugfixes
